fix: move to astro-iconify (#11)
* chore(deps): update

* fix: move to astro-iconify
This change aims to fix the build issue that occurs when one uses the astro-icon package.
astro-iconify is a more up-to-date version of the astro-icon package that fixes the svgo issue, as well as bringing other dependencies and etc.
Note that this change is only temporary until astro-icon is updated.